yearbook_ccs/requirements.txt
six - for the tokens.py under user_management apps
python-dotenv - to access envi variables
django-imagekit  